requirements.txt
psycopg2-binary
djangorestframework
django-cors-headers
pyyaml
uritemplate
python-dotenv
